Widen animation_update_internal render/game-state parameter to long
January reads and forwards parameter 1 as a dword; the header owner and definition now agree. Full build: +1 exact (304 padded), 0 regressions, parks clean.

+++ source/models/model_animations.c
@@ -435,7 +435,7 @@ void animation_set_frame_size(
 }
 
 short animation_update_internal(
-	short render_or_affects_game_state,
+	long render_or_affects_game_state,
 	long animation_graph_index,
 	struct animation_state *state,
 	long *sound_index)
